+++ NEWS.md
@@ -5,6 +5,7 @@
 ## mltools 0.1.0.9000
 
 - Added `date_factor()` for converting dates to a grouped ordered factor (e.g. months, yearmonths, yearquarters, etc.)
+- Implemented stratification for numeric `x` in `folds(x, stratified=TRUE, ...)`
 
 ---